Refactor ContainerFactory to extract mixed concerns
Extracted container building logic into focused private methods:
- registerCoreServices(): Loads default config and sets parameters
- loadBridges(): Discovers and loads MCP extension bridges
- registerUserServices(): Loads user-defined service configurations
- loadEnvironmentVariables(): Loads environment variables from .env files

The create() method now orchestrates these four steps with clear
separation of concerns and improved maintainability.

Also fixed bug in environment variable loading where .env.local file
was being added to the load list when it didn't exist (inverted logic).

All 27 existing tests pass, confirming backward compatibility.
